Remove Library
Xoa Thu Vien va chinh lai bang unsigned

+++ ADC.h
@@ -1,33 +1,31 @@
 #ifndef _ADC_H
 #define _ADC_H
-#include <stm32f4xx.h>
-#include <stdint.h>
 #define ADC1_BASE_ 0x40012000UL
 typedef struct
 {
-    volatile uint32_t SR;
-    volatile uint32_t CR1;
-    volatile uint32_t CR2;
-    volatile uint32_t SMPR1;
-    volatile uint32_t SMPR2;
-    volatile uint32_t JOFR1;
-    volatile uint32_t JOFR2;
-    volatile uint32_t JOFR3;
-    volatile uint32_t JOFR4;
-    volatile uint32_t HTR;
-    volatile uint32_t LTR;
-    volatile uint32_t SQR1;
-    volatile uint32_t SQR2;
-    volatile uint32_t SQR3;
-    volatile uint32_t JSQR;
-    volatile uint32_t JDR1;
-    volatile uint32_t JDR2;
-    volatile uint32_t JDR3;
-    volatile uint32_t JDR4;
-    volatile uint32_t DR;
+    volatile unsigned long SR;
+    volatile unsigned long CR1;
+    volatile unsigned long CR2;
+    volatile unsigned long SMPR1;
+    volatile unsigned long SMPR2;
+    volatile unsigned long JOFR1;
+    volatile unsigned long JOFR2;
+    volatile unsigned long JOFR3;
+    volatile unsigned long JOFR4;
+    volatile unsigned long HTR;
+    volatile unsigned long LTR;
+    volatile unsigned long SQR1;
+    volatile unsigned long SQR2;
+    volatile unsigned long SQR3;
+    volatile unsigned long JSQR;
+    volatile unsigned long JDR1;
+    volatile unsigned long JDR2;
+    volatile unsigned long JDR3;
+    volatile unsigned long JDR4;
+    volatile unsigned long DR;
 } ADC_TYPE;
 #define ADC1_CONTROL ((ADC_TYPE*) ADC1_BASE_)
 void ADC1_Init(void);
 void ADC1_Start(void);
-void ADC1_Read_All(uint16_t *buffer);
+void ADC1_Read_All(unsigned short *buffer);
 #endif
