+++ DMA.c
@@ -1,7 +1,7 @@
 #include "DMA.h"
 #include "ADC.h"
 
-void DMA2_Stream0_ADC_Config(uint16_t *buffer)
+void DMA2_Stream0_ADC_Config(unsigned short *buffer)
 {
     DMA2_CONTROL->STREAM[0].CR &= ~(1<<0); // disable stream
     while(DMA2_CONTROL->STREAM[0].CR & 1);
@@ -9,8 +9,8 @@ void DMA2_Stream0_ADC_Config(uint16_t *buffer)
     DMA2_CONTROL->LIFCR = 0xFFFFFFFF;
     DMA2_CONTROL->HIFCR = 0xFFFFFFFF;
 	
-    DMA2_CONTROL->STREAM[0].PAR  = (uint32_t)&ADC1_CONTROL->DR; // Chon dia chi nguon du lieu
-    DMA2_CONTROL->STREAM[0].M0AR = (uint32_t)buffer; // Chon dia chi noi luu du lieu
+    DMA2_CONTROL->STREAM[0].PAR  = (unsigned long)&ADC1_CONTROL->DR; // Chon dia chi nguon du lieu
+    DMA2_CONTROL->STREAM[0].M0AR = (unsigned long)buffer; // Chon dia chi noi luu du lieu
 
     DMA2_CONTROL->STREAM[0].NDTR = 4; // So luong du lieu can transfer
